--- a/std/stdio.d
+++ b/std/stdio.d
@@ -1560,6 +1560,45 @@ is recommended if you want to process a complete file.
         }
     }
 
+    unittest
+    {
+        static import std.file;
+        auto deleteme = testFilename();
+        std.file.write(deleteme, "123\n456789");
+        scope(exit) std.file.remove(deleteme);
+
+        auto file = File(deleteme);
+        char[] buffer = new char[10];
+        char[] line = buffer;
+        file.readln(line);
+        auto beyond = line.length;
+        buffer[beyond] = 'a';
+        file.readln(line); // should not write buffer beyond line
+        assert(buffer[beyond] == 'a');
+    }
+
+    unittest // bugzilla 15293
+    {
+        static import std.file;
+        auto deleteme = testFilename();
+        std.file.write(deleteme, "a\n\naa");
+        scope(exit) std.file.remove(deleteme);
+
+        auto file = File(deleteme);
+        char[] buffer;
+        char[] line;
+
+        file.readln(buffer, '\n');
+
+        line = buffer;
+        file.readln(line, '\n');
+
+        line = buffer;
+        file.readln(line, '\n');
+
+        assert(line[0 .. 1].capacity == 0);
+    }
+
 /** ditto */
     size_t readln(C, R)(ref C[] buf, R terminator)
     if (isSomeChar!C && is(Unqual!C == C) && !is(C == enum) &&
@@ -4043,35 +4082,42 @@ private struct ReadlnAppender
     import core.stdc.string;
 
     char[] buf;
-    size_t cap;
     size_t pos;
-    size_t initcap;
+    bool safeAppend = false;
 
     void initialize(char[] b)
     {
         buf = b;
-        cap = initcap = buf.capacity;
-        if (cap < buf.length)
-            cap = buf.length;
         pos = 0;
     }
     @property char[] data()
     {
-        // always shrink/extend the buffer to allow reuse in the next call
-        if (initcap > 0 || cap > buf.length)
+        if (safeAppend)
             assumeSafeAppend(buf.ptr[0..pos]);
         return buf.ptr[0..pos];
     }
 
     void reserve(size_t n)
     {
-        if (pos + n > cap)
+        if (buf.length >= pos + n) // buf is already large enough
+            return;
+
+        immutable curCap = buf.capacity;
+        if (curCap >= pos + n)
         {
-            size_t ncap = cap * 2 + 128 + n;
+            buf.length = curCap;
+            /* Any extra capacity we end up not using can safely be claimed
+            by someone else. */
+            safeAppend = true;
+        }
+        else
+        {
+            size_t ncap = buf.length * 2 + 128 + n;
             char[] nbuf = new char[ncap];
             memcpy(nbuf.ptr, buf.ptr, pos);
-            cap = nbuf.capacity;
-            buf = nbuf.ptr[0 .. buf.length];   // remember initial length
+            buf = nbuf;
+            // Allocated a new buffer. No one else knows about it.
+            safeAppend = true;
         }
     }
     void putchar(char c)
@@ -4097,17 +4143,17 @@ private struct ReadlnAppender
     }
     void putonly(char[] b)
     {
+        import std.algorithm: max;
         assert(pos == 0);   // assume this is the only put call
-        if (b.length > cap)
+        if (b.length > max(buf.length, buf.capacity))
         {
             buf = b.dup;
-            initcap = 0; // no need to call assumeSafeAppend
+            pos = b.length;
         }
         else
         {
-            memcpy(buf.ptr, b.ptr, b.length);
+            putbuf(b);
         }
-        pos = b.length;
     }
 }