@schveiguy 2015-11-12
ping @rainers 
I don't think this is the right solution, because I don't agree with the unit tests, but I can't get the error to happen on my macOS system, so I'm having trouble diagnosing the issue.