@aG0aep6G 2015-11-12
> I can't get the error to happen on my macOS system, so I'm having trouble diagnosing the issue.

`ReadlnAppender` is only used in `version (DIGITAL_MARS_STDIO)` and `version (MICROSOFT_STDIO)`. OS X uses the `version (HAS_GETDELIM)` variant.