@schveiguy 2015-11-12
> ReadlnAppender is only used in [Windows]

Ah, ok. I forgot that.