@schveiguy 2015-11-12
> I don't think this is the right solution

I'm having second thoughts on this opinion. I didn't realize that the other versions (non ReadlnAppender) didn't care about the returned buffer's appending status.

I'm thinking that this update (after issues noted above) will be acceptable, because readln isn't meant for looping over (and when it is, you can use the same mechanism as byLine, which should be high performance).