@aG0aep6G 2015-11-12
Amended fixes for the noted issues.