@aG0aep6G 2015-11-12
Amended the `newBuf` thing. Calling it `safeAppend` now, though. Thanks @schveiguy