@schveiguy 2015-11-12
LGTM. I think this is actually going to end up making `byLine` faster too, because `assumeSafeAppend`/`capacity` is only called on expansion (those need to dig into the GC metadata which can be expensive).

@rainers please review, this changes the implications of `readln`, but I think for the better.