@rainers 2015-11-12
I agree the test case is valid. With a slightly more elaborate GC, that reclaims memory "freed" by assumeSafeAppend, it might reuse memory still referenced by other slices.

It just took 8 month for https://github.com/D-Programming-Language/phobos/pull/2794 to be merged, and it was a more obvious memory corruption. It wasn't me who insisted on performance over correctness, but that was a concern to others, so you might want to check the benchmark given there to see the actual change in performance by this PR.