@aG0aep6G 2015-11-12
> you might want to check the benchmark given there to see the actual change in performance by this PR.

no optimization flags:
readln1: 129 ms -> 158 ms
readln2: 148 ms -> 124 ms
byLine: 178 ms -> 161 ms

-release -O -inline:
readln1: 94 ms -> 125 ms
readln2: 110 ms -> 90 ms
byLine: 87 ms -> 77 ms

Base is 2.069.0. Tested in wine, not proper Windows.