@rainers 2015-11-13
Here are some numbers from my Windows system
```
GIT head                                              this PR
win32:
readln1: 117 ms                                       readln1: 139 ms
readln2: 133 ms                                       readln2: 94 ms
byLine: 104 ms                                        byLine: 86 ms
2192 bytes used, 1048576 bytes pool mem               941040 bytes used, 1048576 bytes pool mem
GC summary:    1 MB,    1 GC    0 ms                  GC summary:    1 MB,   10 GC    4 ms

win64:
readln1: 333 ms                                       readln1: 397 ms
readln2: 341 ms                                       readln2: 335 ms
byLine: 320 ms                                        byLine: 329 ms
5776 bytes used, 1048576 bytes pool mem               970112 bytes used, 1048576 bytes pool mem
GC summary:    1 MB,    1 GC    0 ms                  GC summary:    1 MB,   27 GC    4 ms

win32mscoff:
readln1: 277 ms                                       readln1: 355 ms
readln2: 290 ms                                       readln2: 287 ms
byLine: 263 ms                                        byLine: 278 ms
5328 bytes used, 1048576 bytes pool mem               178128 bytes used, 1048576 bytes pool mem
GC summary:    1 MB,    1 GC    0 ms                  GC summary:    1 MB,   28 GC    6 ms
```
GC activity happens in the readln1 test case only. We can estimate the number of actually allocated memory by multiplying the number of GC runs (-1 for the final run by the runtime) by the size of the pool (1MB). My test file has a size of 14 MB.