@schveiguy 2015-11-13
1. I think readln1 test is not optimal, and we are discouraging that method. Since `byLine` does something better and we have documented that, I wouldn't put any stock into the lower performance there.
2. readln2 test can be made more optimal:

```D
void testReadln2()
{
    auto f = File(filename);
    size_t n = 0;

    StopWatch sw;
    sw.start();
    char storage[200];
    char buf = storage[]
    while(!f.eof())
    {
        char[] ln = buf;
        f.readln(ln);
        n += ln.length;
        if(ln.length > buf.length) buf = ln;
    }
    auto dur = sw.peek();
    writefln("readln2: %d ms", dur.to!("msecs", int));
}
```

I agree with the changes made, still LGTM.