@schveiguy 2015-11-13
> but that was a concern to others

Note that the most obvious problem was `byLine` performance, which is now a moot point.