@rainers 2015-11-13
> readln2 test can be made more optimal

Sure, but you have to know pretty much the internals of readln to do it. The test cases are rather trying the "obvious" approaches which were considered necessary to be fast.
As we have even added an example to the documentation how to get the performance of byLine, I don't object merging this PR. The DigitalMars runtime version even benefits for the more common cases, while the Microsoft runtime version has never been optimized for performance anyway.