@schveiguy 2015-11-13
> Sure, but you have to know pretty much the internals of readln to do it.

Not really. If `readln` gives you a bigger buffer, then why wouldn't you prefer it over the stack-allocated smaller one?

Seems like we agree it's worth merging, so I'll do so.