@schveiguy 2015-11-13
Auto-merge toggled on